// packages/shared/src/decision-model-schema.ts
export const decisionModelSchema = {
// ...
  $defs: {
// ...
    "custom-node-content": {
      oneOf: [
        {
          type: "object",
          required: ["kind", "config"],
          properties: {
            kind: {
              const: "jsonataDecisionTable",
            },
            config: {
              $ref: "#/$defs/jsonata-decision-table",
            },
          },
        },
// ...
    },
// ...
  },
};